Add authentication and implemented employee positioning
At the moment just one user is authorized to use kidip. In order to log in
to have to enter the proper credentials in /login. To logout you have to
go to /failed manuelly.
Each employee has got a position number. This number can be changed in the
New and Edit dialog. If list all employees in /employee or when go to the planning page
/planning the employees are shown accroding to their position-number.